[css-filters] Migrate css3/filters/effect-reference-delete.html to WPT
Migrate this test out of third_party/blink/web_tests/css3/filters
and into the WPT-specific directory, adding links to the relevant
specs and a test assertion describing its purpose.

In addition, this CL also adds a new waitForAtLeastOneFrame() helper
function and into the new wpt/common/rendering-utils.js file, since
that logic will be required to migrate the many tests in Chromium
that currently rely on the runAfterLayoutAndPaint() helper function,
from run-after-layout-and-paint.js.

Last, this CL also adapts background-image-blur-repaint.html, which was
recently migrated to WPT, to use this new helper function instead of
implementing the double-chained requestAnimationFrame() calls in the
test file itself.

Bug: 1063749
Change-Id: I29c7f17186c60b6f51bdd7ee66945cefd28e5226