[LayoutNG] Don't set the SetIsPushedByFloats() flag on lineboxes.
If we had an empty linebox with a float inside it (and the float was
shifted due to other floats) it was possible for that linebox to fail
the DCHECK in:
ng_block_layout_algorithm.cc:1584
DCHECK(!normal_child_had_clearance || !is_self_collapsing);

Both of these conditions were true.

This patch changes so that we don't set the IsPushedByFloats flag on
lineboxes.

This changes one fragmentation test, which appears(?) to be incorrect or
debate-able.

We now have the same rendering of the mutlicol test as EdgeHTML did
previously.

Bug: 1064987
Change-Id: Ie326437a637cc57d1b17f91ac5e49847fd13c983