SVG: Increase hit-testing precision on strokes
Skia takes a precision value when generating strokes as an optimization.
When rendering, we pass a value derived from the scale-factor in the
CTM, but when hit-testing we were passing a hard-coded value.  This
worked reasonably well for solid strokes, which aren't impacted much by
the precision.  However, for dashed strokes at sufficiently high scales,
the stroke generated for hit-testing was noticeably different between
hit-testing and rendering.

This change passes the CTM through to the hit-testing code, so that we
can increase the precision used for hit testing by the scale factor (for
dashed strokes).

Bug: 964614
Change-Id: I8dfcb3ade86b9e7b63c7415baab5814ac2286ce8