Fix test failures of flex-one-sets-flex-basis-to-zero-px.html in Firefox
The failed checks[1] are all essentially measuring the height of
a single line of Ahem text, but the original test never setting
the line-height css property, so Firefox gets the default value
which gives a browser-defined behavior. So the fix this in Firefox,
the test can simply set 'line-height: 1' to '.flexbox > div {}'.
In the end, this patch changes the 3 lines in '.flexbox > div {}'
to 'font: 14px/1 Ahem;'.

[1] https://wpt.fyi/results/css/css-flexbox/flex-one-sets-flex-basis-to-zero-px.html?diff&filter=ADC&run_id=455370002&run_id=445640004.

Bug: 1063749
Change-Id: If27a2a50ed514e42b90aea85122116961d584c6e